pyproject.toml
[project]
name = "ksec"
version = "0.4.0"
description = "Kubernetes secrets pretty printer"
authors = [ { name = "Tucker Beck", email = "tucker.beck@gmail.com" } ]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/dusktreader/ksec"
repository = "https://github.com/dusktreader/ksec"
requires-python = ">=3.12"
dependencies = [
    "auto-name-enum>=2.0.0",
    "thefuzz>=0.22.1",
    "typer>=0.12.5",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "ruff>=0.7.2",
    "types-pyyaml>=6.0.12.20240917",
]

[project.scripts]
ksec = "ksec.main:app"

[project.optional-dependencies]
yaml = [
    "pyyaml>=6.0.2",
]

[tool.ruff]
line-length = 100

[tool.mypy]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [ "thefuzz" ]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
